Improve Check-in service integration guide (#638)
This pull request aims to improve the integration documentation for EGI
Check-in. The changes include providing additional clarity regarding the
purpose and differences of the development environment compared to the
demo and production environments.

# Summary

A note has been added to highlight that the development environment is
intended for testing the latest features and may differ from demo/prod
in terms of stability and functionality. These changes aim to improve
the understanding of the integration process and help users make
informed decisions about the integration environment when testing their
Service Providers.

---------

Co-authored-by: Enol Fernández <enol.fernandez@egi.eu>