Remove references to comanage (#712)
<!--
A good PR should describe what benefit this brings to the repository.
This description of the PR will be used as the commit message when
merging it.
Ideally, there is an existing issue which the PR address.

Please check the [Contributing
guide](https://docs.egi.eu/about/contributing/)
for style requirements and advice.
-->

# Summary

<!-- Describe in plain English what this PR does. -->
Still misses links to 
* ~dmcc~
* ggus supporter
And check-in docs are not touched as they are being updated on #710

---

<!-- Add, if any, the related issue here, e.g. #6 -->

**Related issue :**

---------

Co-authored-by: Bruce Becker <bruce.becker@egi.eu>