ci: add periodic link check lychee (#733)
# Summary

This updates the pre-commit hooks and adds a periodic link check against
a set of urls.

It deprecates the use of markdownlinkchecker in favour of lychee.

- [x] First, we add the a job be run daily which uses lychee to check
the links present in a series of pages under a given root.
- [x] Next, we remove the existing link checker periodic actions.

Subsequent PRs can enable opening issues with results of broken links.

Other changes:

- remove check-spelling action: conflcits with cspell, not available in
megalinter via pre-commit, is a source of unnecessary toil.
- Remove links to community.egi.eu to show that lychee is actually doing
its job
- Remove unnused packages in `package.json` which seemed to be included
for potential testing activities which were never done. This addresses
several critical vulnerabilities identified by github (but which we are
not actually exposed by).

---

**Related issue :**

Closed #157

---------

Signed-off-by: Bruce Becker <bruce.becker@egi.eu>