ci: remove dependency on build job (#753)
# Summary

Workflow actions require at least one job with no dependencies. This
workflow was declared to depend on the job "build pr preview" in another
workflow, but that does not work for github actions. For now, the
dependency has been removed, but in principle the build and deploy
workflows should be combined.

**Related issue :** #751