fix: add missing /users/ path in the vault HOME (#764)
Update `VAULT_HOME` to fix the path to the secrets.

I also had issues due to the fact that my `egi` account in my
`oidc-agent` was not having the `voperson_id` scope, thus the `jq` call
was not returning it.
The `voperson_id` scope is not mentioned in the upstream `oidc-agent`
configuration for the EGI provider at
https://indigo-dc.gitbook.io/oidc-agent/user/oidc-gen/provider/egi, that
is linked from
https://docs.egi.eu/users/getting-started/cli/#authentication.
And https://docs.egi.eu/users/aai/check-in/obtaining-tokens/oidc-agent/
is not mentioning it too.