Update training infrastructure page (#770)
<!--
A good PR should describe what benefit this brings to the repository.
This description of the PR will be used as the commit message when
merging it.
Ideally, there is an existing issue which the PR address.

Please check the [Contributing
guide](https://docs.egi.eu/about/contributing/)
for style requirements and advice.
-->

# Summary

Update training infrastructure page with updated information.

---

<!-- Add, if any, the related issue here, e.g. #6 -->

**Related issue :**